+++ package.json
@@ -56,6 +56,7 @@
     "gulp-tag-version": "1.x.x",
     "gulp-util": "3.x.x",
     "gulp-zip": "^4.x.x",
+    "json-loader": "^0.5.7",
     "require-dir": "1.x.x",
     "standard": "^11.0.1",
     "ts-loader": "^4.4.2",
